Module Calculator.Print

val get_eqs_and_ineqs : VPL.t -> Pol.Cs.t list * Pol.Cs.t list

val cs_string : string list -> string

val eq_string : string list -> string

val ineq_string : string list -> string

val sage_beginning_string : string

val var_list_to_parameters : string list -> string

val getVars : VPL.t -> string list

val occurIn : string -> string list -> bool

val filter_vars : string list -> string list -> string list

filter_vars l1 l2 filters l1 so that no element of l2 remains in l1.

val prepare_p : VPL.t -> string list -> VPL.t

prepare_p p sl prepares polyhedron p for plotting, projecting variables from sl if necessary.

val sage_ending_string : int -> string

If the polyhedron is plotted on one or two dimensions, the image is saved to be shown after Sage has terminated. Otherwise, it is directly plotted, as Jmol still works when Sage terminates.

val cs_list_to_string_list : Pol.Cs.t list -> string list

Translates a list of constraints into Sage constraints strings.

val to_plot_custom : string list -> string list -> string -> int -> string

val to_plot_for_show : ?⁠vars:string -> VPL.t -> string * int

val to_plot : VPL.t -> string

Gives the Sage string corresponding to the given polyhedron.

val write_file : string -> string -> unit

val append_file : string -> string -> unit