Module Vpl.CstrPoly

module Cs = Cstr.Rat

module Poly : Poly.Type with module Vec = Vector.Rat

type t = {

typ : Cstr_type.cmpT;
the comparison operator

p : Poly.t;
the polynomial

}

A polynomial constraint of the form p typ 0.

val to_string : t -> string

val mk : Cstr_type.cmpT -> Poly.t -> t

mk typ poly builds the linear constraint p typ 0.

val eq : Poly.t -> t

eq p builds the equality p = 0

val le : Poly.t -> t

le p builds the equality p <= 0

val lt : Poly.t -> t

lt p builds the equality p < 0

val compl : t -> t

val empty : t

empty represents 0 = 0

val equal : t -> t -> bool

equal c1 c2 checks that the two polynomials within constraints are syntactically equal.

val cmp : t -> t -> int

syntaxic comparison

val compare : t -> t -> int

val toCstr : t -> Cs.t

val ofCstr : Cs.t -> t

val partition_affine : t list -> Cs.t list * t list

Separates the affine constraints from the given list of polynomial constraints.