Module Vpl.Cstr_type

type cmpT =

| Eq

| Le

| Lt

Type of comparators

type cmpT_extended =

| EQ

| LE

| LT

| GE

| GT

| NEQ

Extended type of comparators.

exception BadMult

See mulc.

exception CannotElim

Exception that may be raised when trying to eliminate a variable

exception NoElim

Exception that may be raised when trying to eliminate a variable

module type Type = sig ... end

Interface of affine constraints.