Module FactoryMaker.Make

Functor that builds a factory from a factory maker.

Parameters

F : Type

Signature

include Type

type cert

Type of certificates.

val factory : cert Factory.t

The factory itself.

val mk : Cstr.Rat.t -> cert

Builds a certificate from a contraint.

val equal : Cstr.Rat.t -> cert -> bool

Checks if a constraint and a certificate represent the same space.

val mkCons : Cstr.Rat.t -> cert Cons.t

Builds a cons from a constraint.

val convert : 'c Pol.t -> cert Pol.t

Converts polymorphic certificates of a polyhedron into type F.cert.

val check : cert Pol.t -> bool

Checks if all certificates of a polyhedron correspond to their associated contraints.

val to_string : cert -> string

Pretty-printer for certificates.