Module HLP.Bound

val getWitness : Var.t list -> HOtypes.LPMaps.t -> Var.t -> Splx.t -> Splx.witness_t option

get vars typ v sx returns a witness allowing to compute the tightest (upper or lower depending on typ bound of v in the polyhedron represented by sx. It returns None if such bound does not exists (i.e. is the polyhedron is unbounded in that direction).

val getBoundIndex : Splx.witness_t -> HOtypes.Poly.t list -> Hi.boundIndex

getBoundIndex wit ph returns the type Hi.boundIndex corresponding to wit, which depends on the order of the constraints in ph.

val get : Var.t list -> HOtypes.LPMaps.t -> Var.t -> Splx.t -> HOtypes.Poly.t list -> Hi.boundIndex option

val updateMaps : Var.t list -> (Var.t * HOtypes.LPMaps.t) list -> Splx.t -> HOtypes.LPMaps.mapDetBound -> HOtypes.LPMaps.mapBound -> HOtypes.Poly.t list -> HOtypes.LPMaps.mapDetBound * HOtypes.LPMaps.mapBound * Hi.boundIndex list option