Module Vpl.HOtypes

module CP = CstrPoly

module Poly = CP.Poly

module Debug : DebugTypes.Type

module MapP : Stdlib.Map.S with type MapP.key = Poly.t

This map binds things to elements of type Poly.t.

val neg_poly : CstrPoly.t -> Poly.t list

module MapPolyHi : sig ... end

Module associating in a map a list of Hi.t to a polynomial. It is used to remember what products were used to cancel a given polynomial.

module MapI = IndexBuild.MapI

This map binds things to elements of type Index.Int.t.

module MapIndexP : sig ... end

This module associates in Handelman indexes to polynomials.

module MapV : Stdlib.Map.S with type MapV.key = Var.t

This map binds things to elements of type Var.t.

module LPMaps : sig ... end

This module defines the maps used in HLP. They gather information about variable bounds.