Module HOtypes.MapIndexP

This module associates in Handelman indexes to polynomials.

type t = Poly.t MapI.t

val of_polyhedron : Poly.t list -> t

Initializes a map from a list of constraints.

val to_string : t -> string

Conversion into string.

val poly_to_deg_max : Poly.t -> Var.t list -> Index.Int.t

returns

an index ind whose ith component is the maximum power in p of the ith variable in l. For instance,

poly_to_deg_max (x^2y - 2x^3 + y^2 + xz^2) [x;z] = (3,2)

poly_to_dex_max (x^2z - z^3) [x;y;z] = (2,0,3)

parameter p

the polynomial

parameter vl

the list of variables

val get : Index.Int.t -> t -> IndexBuild.Map.t -> Poly.t * t * IndexBuild.Map.t

get ind mapIP mapI returns the polynomial corresponding to index ind in map mapIP. If ind has no binding in mapIP, maps are updated.