Module Domain_T.Poly

Type fo polynomials.

include Poly.Type

module Vec : Vector.Type

module Coeff = Vec.Coeff

Type of coefficients.

type exp = int

Type of exponent: a nonnegative integer.

module MonomialBasis : sig ... end

MonomialBasis represents the variables of a monomial, i.e. without the coefficient.

module Monomial : sig ... end

Module of monomials.

type t = Monomial.t list

Type of polynomials.

val to_string : t -> string

Pretty-printer for polynomials.

val mk : Monomial.t list -> t

Builds a polynomial from a list of Monomial.t.

val mk_list : ((Var.t * int) list * Coeff.t) list -> t

Builds a polynomial from a list of monomials.

val fromVar : Var.t -> t

Builds a polynomial from a single variable.

val cste : Coeff.t -> t

Builds a polynomial from a constant.

val ofCstr : Vec.t -> Coeff.t -> t

Builds the polynomial form a constraints.

parameter vec

the linear part of the constraint

parameter cste

the constant part of the constraint

returns

the polynomial vec + coeff

val toCstr : t -> Vec.t * Coeff.t

returns

the linear part and the coefficient of the polynomial.

raises Invalid_argument

if the polynomial is not linear.

val of_string : string -> t

Builds a polynomial from a string.

val data : t -> Monomial.t list

returns

the polynomial p as a Monomial.t list

val compare : t -> t -> int

Comparison between two polynomials. If monomials are equal, compares over coefficients. See MonomialBasis.compare.

val z : t

Constant polynomial equal to zero.

val u : t

Constant polynomial equal to one.

val negU : t

Constant polynomial equal to minus_one.

val is_constant : t -> bool

returns

true if the polynomial is constant

val isZ : t -> bool

returns

true if the polynomial is zero.

val is_linear : t -> bool

returns

true if the polynomial is linear (i.e. degre <= 1). The input polynomial is assumed to be in canonical form.

val equal : t -> t -> bool

Equality test between two polynomials.

val add_cste : t -> Coeff.t -> t

Adds a constant to a polynomial.

val add : t -> t -> t

Addition of two polynomials.

val mul : t -> t -> t

Product of two polynomials.

val mulc : t -> Coeff.t -> t

Product of a polynomial by a constant.

val div : t -> t -> t

Division of a polynomial by another.

raises Div_by_non_constant

if p2 is non-constant

val neg : t -> t

returns

the negation of the polynomial

val sub : t -> t -> t

Substraction of two polynomials.

val sub_monomial : t -> MonomialBasis.t -> t

Removes a monomial from a polynomial.

parameter the

polynomial

parameter the

monomial basis to remove The coefficient is not watched, only the variables.

val sum : t list -> t

Sum a list of polynomials.

val prod : t list -> t

Product of a list of polynomials.

val pow : t -> int -> t

Power of a polynomial by an integer.

val get_max_exponent : Var.t -> t -> int

returns

the maximum exponent of the variable in the polynomial.

val rename : t -> Var.t -> Var.t -> t

Renames a variable in the polynomial.

parameter p

the polynomial

parameter fromX

the variable to rename

parameter toY

the new variable name

val change_variable : (MonomialBasis.t -> MonomialBasis.t option) -> t -> t

Applies the given change of variables on the polynomial.

val monomial_coefficient : t -> MonomialBasis.t -> Coeff.t

returns

the scalar coefficient of the given monomialBasis in the polynomial.

val monomial_coefficient_poly : t -> MonomialBasis.t -> t

returns

the polynomial coefficient of the given monomialBasis in p. For instance, monomial_coefficient_poly 3x1x2x3 - 2x2x3 {ul [x2;x3]} = 3x1 - 2

val get_constant : t -> Coeff.t

returns

the constant coefficient of the polynomial.

val get_linear_part : t -> Var.t list -> t

returns

the linear part of the polynomial. It means all terms that are linear according to is_linear.

val get_vars : t -> Vpl.Var.Set.t

returns

the set of variables of the polynomial.

val horizon : t list -> Var.t

returns

the next unbounded variable in the list of polynomials.

val eval : t -> (Var.t -> Coeff.t) -> Coeff.t

Evaluates a polynomial by replacing each variable with its value given by a function.

parameter p

the polynomial to evaluate

parameter f

the evaluation function

val eval_partial : t -> (Var.t -> Coeff.t option) -> t

Evaluates a polynomial with a partial evaluation function. If a variable has no value in f, it remains in the result.

parameter p

the polynomial to evaluate

parameter f

the partial evaluation function

val gradient : t -> t Rtree.t

Computes the gradient of the polynomial. The result if a map associating to each variable its partial derivative.

val coqvar_to_var : ProgVar.PVar.t -> Var.t

val var_to_coqvar : Var.t -> ProgVar.PVar.t

val n_to_coeff : N.t -> Vec.Coeff.t

val coeff_to_n : Vec.Coeff.t -> N.t