Module Index.Make

Functor of Index

Parameters

Coeff : Scalar

Signature

type coeff = Coeff.t

Type of coefficients

type t = coeff list

Type of Index

val init : int -> t

init n returns an index of length n filled with zeros.

val len : t -> int

len ind returns the length of index ind.

val is_nonnegative : t -> bool

is_nonnegative ind returns true if every element of index ind is nonnegative.

val well_formed : t -> bool

well_formed ind returns true if ind is well formed, false otherwise.

val mk : coeff list -> t

mk l builds an index from the coefficient list l.

raises Invalid_argument

if l contains negative elements.

val unitary : int -> int -> t

unitary i len returns an index of length len, whose value is 1 at index i and 0 otherwise.

val sum : t -> Coeff.t

sum ind returns the sum of all elements of ind.

val set : t -> int -> Coeff.t -> t

set ind i c returns an index equal to ind, with the ith coefficient equals to c.

val get : t -> int -> Coeff.t

get ind i returns the ith coefficient of index ind.

val incr : t -> int -> t

incr ind i returns and index equal to ind such that ith coefficient equals to ind(i)+1.

val decr : t -> int -> t

decr ind i returns and index equal to ind such that ith coefficient equals to ind(i)-1.

val add : t -> t -> t

add ind1 ind2 returns an index such that the ith coefficient equals to ind1(i) + ind2(i).

val sumI : t list -> t

sumI il computes the sum of all elements in il.

raises Invalid_argument

if l is empty.

val sub : t -> t -> t

sub ind1 ind2 returns an index such that the ith coefficient equals to ind1(i) - ind2(i).

val equal : t -> t -> bool

equal ind1 ind2 returns true if for all i ind1(i) = ind2(i), false otherwise.

val is_null : t -> bool

is_null ind returns true if for all i ind(i) = 0, false otherwise.

val first_positive : t -> int

first_positive ind returns the place of the first strictly positive coefficient.

raises Not_found

if there is no such coefficient.

val to_string : t -> string

val compare : t -> t -> int

compare ind1 ind2 returns a positive integer if ind1 > ind2 (w.r.t. lexicographic order), a negative one if ind1 < ind2, 0 otherwise.

val le : t -> t -> bool

le ind1 ind2 returns true if ind1 <= ind2 (w.r.t elementwise order), false otherwise.

val le_nl : t -> t -> bool

le_nl ind1 ind2 returns true if ind1 <= ind2 (w.r.t elementwise order), false otherwise. It ignores coefficients that are equal to one.

val is_unitary : t -> bool

is_unitary ind returns true if ind contains a single one, and every other coefficient is null, false otherwise.

val one_coeff_nn : t -> bool

one_coeff_nn ind returns true if ind contains a single nonnull coefficient, false otherwise.

val data : t -> coeff list