Module IndexBuild.IndexList

type t = Index.Int.t list

val to_string : t -> string

val equal : t -> t -> bool

val get_next : int -> t -> Index.Int.t

get_next dim l returns an index that is either an element of l or an index smaller than max 2 (l/2) elements of l. In the second case, this index is useful to compute l efficiently. If l is a singleton, get_next returns a null index.

raises Failure

if no such element has been found.

val components : Index.Int.t -> t

compute ind returns the list of nonnull components of ind. For instance, compute [1;0;2] = [[1;0;0] ; [0,0,2]].

val get_preds : Index.Int.t -> t

get_preds ind returns the list of indexes that are strictly smaller than ind w.r.t function dunction Index.Int.le.

val le : int -> int -> t

le dim ind_max returns the list of indexes of length dim whose value is smaller than ind_max w.r.t Index.Int.value.