Module Vpl.IneqSet

module Debug : DebugTypes.Type

type 'c t = {

ineqs : 'c Cons.t list;
Inequalities of the polyhedron

regions : 'c PLP.Region.t list option;
Regions of a polyhedron

}

Type of inequality set.

val top : 'c t

val is_top : 'c t -> bool

returns

true if the polyhedron is top.

val of_list : 'c Cons.t list -> 'c t

val to_string_raw : 'c t -> string

Conversion into string.

val to_string : (Var.t -> string) -> 'c t -> string

Conversion into string.-

parameter varPr

the variable printer

val to_string_ext : 'c Factory.t -> (Var.t -> string) -> 'c t -> string

Conversion into string. Also prints the certificates.

parameter factory

the factory

parameter varPr

the variable printer

val map : ('c1 -> 'c2) -> 'c1 t -> 'c2 t

Applies a function to each certificate.

parameter f

the funciton to apply

type 'c prop_t =

| Empty of 'c

| Trivial

| Implied of 'c

| Check of 'c Cons.t

Result of a syntactic inclusion test. Certificates are the constraints that must be added to obtain the associated result.

val synIncl : 'c Factory.t -> 'c EqSet.t -> 'c t -> Cstr.Rat.t -> 'c prop_t

Checks if a constraint includes an inequality set. The constraint is first rewritten with an equality set.

parameter factory

the factory

parameter es

an equality set

parameter s

an inequality set

parameter cstr

a constraint

type 'c rel_t =

| NoIncl

| Incl of 'c list

Result of an inclusion test.

val incl : 'c1 Factory.t -> Var.t -> 'c1 EqSet.t -> 'c1 t -> 'c2 t -> 'c1 rel_t

Checks the inclusion between two inequality sets. The constraint is first rewritten with an equality set.

parameter factory

the factory

parameter the

horizon variable

parameter es

an equality set

parameter s1

an inequality set

parameter s2

an inequality set Tests if s2 includes s1.

val rename : 'c Factory.t -> 'c t -> Var.t -> Var.t -> 'c t

Variable renaming in an inequality set

parameter factory

the factory

parameter s

an inequality set

parameter fromX

variable to rename

parameter toY

new variable name

val assume : Var.t -> 'c t -> 'c Cons.t list -> Scalar.Symbolic.t Rtree.t -> 'c t

Assume a list of constraints.

parameter h

horizon variable

parameter s

the initial inequality set

parameter conss

the list of constraints to assume

parameter point

a point within the polyhedron's interior

val assume_back : 'c Factory.t -> 'c t -> 'c Cons.t -> Vector.Symbolic.t -> ('c t * Vector.Symbolic.t) option

val subst : 'c Factory.t -> Var.t -> 'c EqSet.t -> Var.t -> 'c Cons.t -> 'c t -> 'c t

Substitutes a variable with an expression in an equality set.

parameter factory

the factory

parameter h

the hoizon variable

parameter es

the equality set

parameter x

variable to substiture

parameter e

the equality that defines x

parameter s

the inequality set to rewrite

val fmElim : 'c Factory.t -> Var.t -> 'c EqSet.t -> Var.t option Rtree.t -> 'c t -> 'c t

Eliminates a variable with Fourier-Motzkin elimination.

parameter factory

the factory

parameter h

the horizon variable

parameter es

the equality set

parameter mask

variables to eliminate

parameter s

the inequality set

val plpElim : 'c Factory.t -> Vector.Rat.t -> Var.t list -> 'c t -> 'c t

Eliminates a a variable by PLP.

parameter factory

the factory

parameter point

the normalization point

parameter vars

variables to eliminate

parameter s

the inequality set

val proj_incl : 'c Factory.t -> Vector.Rat.t -> Var.t list -> 'c EqSet.t -> 'c t -> 'c t -> 'c t option

val get_regions_from_point : 'c Factory.t -> 'c t -> Vector.Rat.t -> ('c Cons.t list * Vector.Symbolic.t) list

val satisfy : 'c t -> Vector.Rat.t -> bool

returns

true if the given point satisfies the constraints.