Module IntervalizationOracle.Oracle

Builds an OracleType from a

Parameters

T : IOtypes.Type

Signature

module P = T.D.Poly

module M = T.D.Poly.Monomial

module MB = T.D.Poly.MonomialBasis

type prophecy = {

env : T.env;
Function from variables to static intervals.

mode : IOtypes.mode;
Side of approximation.

current_mb : MB.t T.MapMonomial.t;
Current state of the monomial

var_to_keep : Var.t option T.MapMonomial.t;
Variable to keep.

map_itv : int T.MapV.t T.MapMonomial.t;
Map from monomials to te number of times they are intervalized in the monomial.

terms : T.D.BasicTerm.term list;
List of terms that will replace the monomial.

}

module type Prayer = sig ... end

val recursive_oracle : (P.t -> prophecy -> (module Prayer) list -> prophecy) Stdlib.ref

val sign_change : M.t -> Var.t -> T.env -> bool

returns

true if the sign of the variable is changed by the monomial.

val get_map_itv : MB.t -> prophecy -> int T.MapV.t

val get_itv_degree : MB.t -> Var.t -> prophecy -> int

val get_var_to_keep : MB.t -> prophecy -> Var.t option

val get_current_mb : MB.t -> prophecy -> MB.t

val set_var_to_keep : MB.t -> Var.t -> prophecy -> prophecy

val update_monomial : MB.t -> prophecy -> MB.t

returns

the current representation of the given monomial, taking into account its intervalized variables.

val set_current_mb : MB.t -> prophecy -> prophecy

val add_itv : prophecy -> MB.t -> Var.t -> int -> prophecy

Adds an intervalization to a variable for a given monomial.

parameter pr

the initial prophecy

parameter mb

the monomial basis

parameter var

the variable th annotate

parameter i

the number of time the variable must be intervalized

returns

the prophecy where var has been annoted with annot in mb

module Multiplicity : Prayer

module LinearMonomial : Prayer

A monomial is linear.

module ConstantMonomial : Prayer

Other prophecies can generate constant monomials.

module VarCste : Prayer

A constant variable can be statically intervalized.

module GreatestInterval : Prayer

All variables are bounded, and we keep the variable with the greatest interval.

module OneUnboundedVar : Prayer

Exactly one variable is unbounded in the monomial.

module FirstUnbounded : Prayer

Keeps the first unbounded variable. Otherwise, take the greatest interval.

module Default : Prayer

Always applicable. Keeps the first variable.

module Focusing : sig ... end

Alternative to Default. Remaining variables are centered onto zero.

module VarChosen : Prayer

When a variable is chosen, skip all prophecies.

val prayers : (module Prayer) list