Module Poly.Monomial

Module of monomials.

type t = MonomialBasis.t * Coeff.t

Type of monomial

val null : t

Zero monomial.

val to_string : t -> string

pretty-printer for monomials.

val compare : t -> t -> int

Comparison between two monomials. If monomials are equal, compares over coefficients. See MonomialBasis.compare.

val equal : t -> t -> bool

Equality test between two monomials.

val mk : MonomialBasis.t -> Coeff.t -> t

Builds a monomial from a monomialBasis and a coefficient.

val mk_list : (Var.t * exp) list -> Coeff.t -> t

Builds a monomial.

val data : t -> MonomialBasis.t * Coeff.t

returns

the monomial.

val get_exponent : Var.t -> t -> int

returns

the exponent of the given variable.

val mul : t -> t -> t

Multiplication of two monomials.

val is_linear : t -> bool

returns

true if the degree of m is at most one. The monomial is assumed to be in canonical form.

val is_constant : t -> bool

returns

true if the degree of m is zero. The monomial is assumed to be in canonical form.

val eval : t -> (Var.t -> Coeff.t) -> Coeff.t

Evaluates a monomial with an evaluation function

parameter m

the monomial to evaluate

parameter f

the evaluation function

val eval_partial : t -> (Var.t -> Coeff.t option) -> t

Evaluates a monomial with a partial evaluation function. If a variable has no value in f, it remains in the result.

parameter m

the monomial to evaluate

parameter f

the partial evaluation function

val get_vars : t -> Vpl.Var.Set.t

returns

the set of variables of the monomial.

val change_variable : (MonomialBasis.t -> MonomialBasis.t option) -> t -> t

Applies the given change of variables on the monomial.