Module type Min.Type

module VecInput : Vector.Type

val minimize : VecInput.t -> Cstr.Rat.t list -> (Cstr.Rat.t * VecInput.t) list

minimize x cstrs removes the redundancies in the list of constraints cstrs.

parameter cstrs

is the list of constraints to minimize. Syntactic redundancies should have been removed previously.

parameter x

is a point that should lie in the interior of the polyhedron defined by cstrs. This function attach to each returned constraint cstr a point that violates only cstr. When possible, this point does not saturate the completentary of cstr.

val minimize_cone : VecInput.t -> Cstr.Rat.t list -> (Cstr.Rat.t * VecInput.t) list

val name : string