Module NamedDomain.Make

Parameters

D : AbstractDomain.Type

Signature

type t = D.t named_of

Type of abstract value.

type var = D.var

Type of variables.

type cert

Type of certificates.

type a_expr = D.a_expr

Type of arithmetic expressions.

type b_expr = D.b_expr

Type of boolean expressions.

val top : t

Unbounded abstract value.

val bottom : t

Empty abstract value.

val is_bottom : t -> bool

returns

true if the given abstract value is empty.

val get_bottom_cert : t -> cert option

returns

a unsatisfiability certificate if the given abstract value is bottom

None otherwise

val get_cstrs : t -> Cstr.Rat.t list

returns

the list of constraints of the abstract value.

val get_b_expr : t -> b_expr

returns

the constraints of the abstract value as a boolean expression.

val get_vars : t -> var list

returns

the list of variables bounded in the abstract value.

val meet : t -> t -> t

returns

the intersection of two abstract values.

val join : t -> t -> t

returns

the join of two abstract values.

val project : var list -> t -> t

Eliminates the given list of variables from a abstract value, by orthogonal projection. This operator relies on a certified single-variable projection. Variables will be eliminated one after the other. For a projection that eliminates several variables at the same time, see project_vars.

val project_vars : var list -> t -> t

Eliminates the given list of variables from a abstract value, by orthogonal projection. Contrary to project, this operator actually eliminates all variables at the same time.

val proj_incl : t -> t -> t option

proj_incl p1 p2 returns

either the projection of p1 on p2's variables

None if the projection of p1 on p2's variables does not contain p2

val widen : t -> t -> t

Widening operator.

val leq : t -> t -> bool

Checks the inclusion of two abstract values.

val to_string : (var -> string) -> t -> string

returns

a string representing the given abstract value. Requires a pretty-printer for variables.

val get_upper_bound : a_expr -> t -> Pol.bndT option

returns

the upper bound of an expression in an abstract value.

val get_lower_bound : a_expr -> t -> Pol.bndT option

returns

the lower bound of an expression in an abstract value.

val itvize : a_expr -> t -> Pol.itvT

returns

both the upper and lower bounds of an expression in a abstract value.

val assume : b_expr -> t -> t

Guard operator.

returns

the intersection between a condition and a abstract value.

val asserts : b_expr -> t -> bool

returns

true if the given condition is satisfied in the given abstract value.

val diff : t -> t -> t list

diff p1 p2

returns

a list of abstract values whose union is p1 \ p2.

val get_regions : t -> t list

returns

a partition into regions of the abstract value.

val split_in_half : t -> t list

returns

the partition of the abstract value into two halfs according to the greatest axis-aligned interval.

val set_point : Vector.Rat.t -> t -> t

Sets up the normalization point used for PLP-based operators.

val size : t -> Scalar.Rat.t option

returns

the size of the given abstract value. Semantic is left unspecified.

val assign : (var * a_expr) list -> t -> t

Computes the effect of parallel assignments on a abstract value.

val spawn : t -> Vector.Rat.t

returns

a point within the abstract value

val satisfy : Vector.Rat.t -> t -> bool

satisfy vec p

returns

true if the point vec belongs the the abstract value p