Module type OracleType.Prayer

Prayer interface.

val name : string

Prayer Name.

type pneuma

Type of elements needed by the prayer.

val kill_when_fail : bool

If set to true, the prayer won't be tried again if it fails once.

val pray : P.t -> prophecy -> pneuma option

Praying for a pneuma to come. This is the parsing step of the strategy. The prayer determines if it can apply on the polynomial.

parameter p

the polynomial to handle

parameter pr

the current prophecy

returns

None if the prayer can not be applied

returns

Some pneuma if the prayer can be applied

val inhale : P.t -> prophecy -> pneuma -> P.t * prophecy

Inhale the pneuma to generate a prophecy. If the prayer returned a pneuma, it is provided to this function that executes the strategy.

parameter p

the polynomial to handle

parameter the

current prophecy

returns

p' the new polynomial to handle

returns

pr' the new prophecy