Module Next_Point.Greedy

val get_next : 'c t -> (int * Boundary.t) -> ExplorationPoint.t option

val get_next_point_rec : region_t -> 'c t -> 'c t

val get_next_point : region_t -> 'c t -> ExplorationPoint.t option * 'c t

Returns the next point to explore, or None if there is none. The returned point is the first element of plp.todo.