Module Next_Point.Point_in_Region

val eval_regions : ExplorationPoint.t -> 'c t -> (int * 'c Region.t) option

Returns the index of the region in which the given point is minimal w.r.t objective functions.

val in_region' : ExplorationPoint.t -> 'c t -> (int * 'c Region.t) option

Returns a region containing the given point, if it exists.

val in_region : ExplorationPoint.t -> 'c t -> (int * 'c Region.t) option