Module PLPCore.Region

type 'c t = {

id : int;

r : (Boundary.t * int option) list;

point : Vec.t;

sx : 'c PSplx.t;

}

val mk : int -> Boundary.t list -> Vec.t -> 'c PSplx.t -> 'c t

val canon : 'c t -> 'c t

val copy : 'c t -> 'c t

val get_cstrs : 'c t -> Cs.t list

val to_string : 'c t -> string

val extract : 'c PSplx.t -> Cs.t list

val getPointInside : region_t -> Var.t -> Cs.t list -> Vec.t option

Returns a point in the regions' interior. The chosen point is a 1 from each boundary if possible.

val contains : 'c t -> Vec.t -> bool

Returns true if the given point lies in the given region.

val contains' : Cs.t list -> Vec.t -> bool

Returns true if the given point lies in the given region.

val contains_large : 'c t -> Vec.t -> bool

Returns true if the given point lies in the given relaxed (i.e. non-strict) region.

val contains'_large : Cs.t list -> Vec.t -> bool

Returns true if the given point lies in the given relaxed (i.e. non-strict) region.

val contains_ep : 'c t -> ExplorationPoint.t -> bool

val contains_large_ep : 'c t -> ExplorationPoint.t -> bool

val equalSyn : 'c t -> 'c t -> bool

val equal : 'c t -> 'c t -> bool

val eval_obj : 'c t -> Cs.Vec.t -> Cs.Vec.Coeff.t option