Module Vpl.PLPIncremental

include PLPCore

module Cs = Cstr.Rat

module Debug = PLPCore.Debug

module Profile = PLPCore.Profile

module Vec = PLPCore.Vec

module Minimization = PLPCore.Minimization

module Explore = PSplxExec.Explore

module Coeff = PLPCore.Coeff

module MapV = PLPCore.MapV

type region_t =

| Cone

| NCone

module Boundary = PLPCore.Boundary

val strict_comp : Cs.t -> Cs.t

val cstr_eq : Cs.t -> Cs.t

val is_trivial : Cs.t -> bool

module ExplorationPoint = PLPCore.ExplorationPoint

module Region = PLPCore.Region

type 'c mapRegs_t = 'c Region.t MapV.t

type 'c t = {

regs : 'c mapRegs_t;

todo : ExplorationPoint.t list;

}

val empty : 'c t

val to_string : 'c t -> string

module Check = PLPCore.Check

Module that checks results on the fly. It is based on Debug.Check, thus these verifications cost almost nothing if checking is disabled.

module Adjacency = PLPCore.Adjacency

module Next_Point = PLPCore.Next_Point

Several methods to choose the next point to explore.

val get_next_point : region_t -> 'a t -> ExplorationPoint.t option * 'a t

val reg_id : int Stdlib.ref

val get_fresh_id : unit -> int

Returns a fresh id for a new region.

module Exec = PLPCore.Exec

Module that runs one exploration

val extract_points : 'c Region.t -> int -> ExplorationPoint.t list

type 'c config = {

add_region : 'c Region.t option -> 'c Region.t -> ExplorationPoint.t -> 'c t -> 'c t;

reg_t : region_t;

points : ExplorationPoint.t list;

regions : 'c Region.t list;

}

module Add_Region = PLPCore.Add_Region

val init_sx : 'c Factory.t -> 'c PSplx.t -> Vec.t -> bool

Leads to a feasible parametric simplex tableau.

val find_new_point : Vec.t -> Vec.t -> Vec.t

This function is used when an execution (ran on point p) led to a region with empty interior. It returns the middle between the initial region's point and p.

val exec : 'c config -> 'c PSplx.t -> 'c t -> 'c t

val result_to_string : ('c Region.t * 'c Cons.t) list option -> string

val get_results : 'c Factory.t -> 'c t -> ('c Region.t * 'c Cons.t) list

Careful: if a region has no simplex tableau (because it was given in input), it does not appear in the result.

val init_regions : 'c config -> 'c PSplx.t -> 'c mapRegs_t

val choose_init_point : 'c PSplx.t -> 'c config -> Vec.t * 'c config

val init_and_exec : 'c Factory.t -> 'c config -> 'c PSplx.t -> 'c t option

val run : 'c Factory.t -> 'c config -> 'c PSplx.t -> ('c Region.t * 'c Cons.t) list option

val fix_tableau : 'c PSplx.t -> unit

val renormalize : Vector.Rat.t -> 'c PSplx.t -> unit

exception EmptyInput

val check_normalization_point : Cs.t list -> 'c Cons.t -> Cs.Vec.t -> Vector.Symbolic.t option

returns

None if the old normalization point is still ok

returns

a new normalization point otherwise

raises EmptyInput

if the new polyhedron is empty

val irredundant : 'a -> Cs.t -> Cs.t list -> bool

val add_column_to_region : 'c Factory.t -> 'c Region.t -> 'c Cons.t -> ('c Region.t * ExplorationPoint.t list) option

val n_deleted : int Stdlib.ref

val n_updated : int Stdlib.ref

val n_new : int Stdlib.ref

val n_total : int Stdlib.ref

val add_column : 'c Factory.t -> 'c config -> Cs.t list -> 'c Region.t list -> 'c Cons.t -> Vector.Symbolic.t -> ('c Region.t * 'c Cons.t) list * Vector.Symbolic.t