Module Vpl.Pol

module Cs = Cstr.Rat

module Vec = Cs.Vec

module Debug : DebugTypes.Type

type 'c t = {

eqs : 'c EqSet.t;

ineqs : 'c IneqSet.t;

point : Vector.Symbolic.t option;

}

The type of an abstract value. The internal representation splits the linear constraints into a set of equalities and a set of inequalities. The vector is a point within the interior of the polyhedron.

val top : 'c t

The polyhedron with no constraints.

val get_eqs : 'c t -> 'c EqSet.t

val get_ineqs : 'c t -> 'c Cons.t list

val get_point : 'c t -> Vector.Symbolic.t

type 'c rel_t =

| NoIncl

| Incl of 'c list

The return type of the inclusion test incl. If inclusion holds, the certificate can be checked using function check.

type bndT =

| Infty

| Open of Scalar.Rat.t

| Closed of Scalar.Rat.t

type itvT = {

low : bndT;

up : bndT;

}

val bnd_to_string : bndT -> string

val get_low : itvT -> bndT

val get_up : itvT -> bndT

val length_itv : itvT -> Scalar.Rat.t option

Returns the length of the given interval, or None if it is unbounded.

type cstT =

| S of Scalar.Rat.t

| I of itvT

type assign_t = {

var : Var.t;

lin : (Scalar.Rat.t * Var.t) list;

cst : cstT;

}

The description of an assignment operation. var is given the value lin + cst.

val getVar : assign_t -> Var.t

val getLin : assign_t -> (Scalar.Rat.t * Var.t) list

val getCst : assign_t -> cstT

val equalSyn : 'c1 t -> 'c2 t -> bool

equalSyn p1 p2 returns true if p1 and p2 are made of exactly the same constraints. The comparison is sensitive to insertion order for equalities, but not for inequalities.

val mk : 'c Factory.t -> 'c Cons.t list -> 'c t option

mk l builds the polyhedron representing the conjunction of constraints in l. If the resulting polyhedron is empty (i.e. the conjunction is unsatisfiable), None is returned.

val mkSub : 'c Factory.t -> Var.t -> 'c Cons.t list -> 'c t option

type 'c meetT =

| Added of 'c t

| Contrad of 'c

A value of type meetT represents the result of an intersection of two sets of constraints s1 and s2. Either this intersection is empty, or it is some new set of constraints s. The constructor Contrad represents the former possibility, where the certificate gives a proof of the intersection being empty. A value Added (s, (fwd, bwd)) states that fwd is the proof of the intersection of s1 and s2 being included in s, while bwd is a proof of s being included in the intersection of s1 and s2.

val meetPr : 'c Factory.t -> (Var.t -> string) -> 'c meetT -> string

Pretty-printer for values of type meetPr.

val meetEq : 'c meetT -> 'c meetT -> bool

Equality t*est for values of type meetT. Certificate equality is tested using Factory.isEq and polyhedron equality is test with isEqSyn.

val add : 'c Factory.t -> 'c t -> 'c Cons.t -> 'c meetT

Intersect a polyhedron with a new constraint.

val addSub : 'c Factory.t -> Var.t -> 'c t -> 'c Cons.t -> 'c meetT

val addM : 'c Factory.t -> 'c t -> 'c Cons.t list -> 'c meetT

Intersect a polyhedron with a list of new constraints.

val addMSub : 'c Factory.t -> Var.t -> 'c t -> 'c Cons.t list -> 'c meetT

val meet : 'c Factory.t -> 'c t -> 'c t -> 'c meetT

Compute the intersection of two polyhedra.

val meetSub : 'c Factory.t -> Var.t -> 'c t -> 'c t -> 'c meetT

val project : 'c Factory.t -> 'c t -> Var.t list -> 'c t

Forget all information about a set of variables. Although the result is identical to projecting the variables one by one, heuristic choices on the projection order make projectM more efficient.

val projectSub : 'c Factory.t -> Var.t -> 'c t -> Var.t list -> 'c t

val proj_incl : 'c Factory.t -> 'c t -> 'c t -> 'c t option

val join : 'c1 Factory.t -> 'c2 Factory.t -> 'c1 t -> 'c2 t -> 'c1 t * 'c2 t

join p1 p2 computes the convex hull p of p1 and p2. The certificate provides the necessary arguments to show inclusion of p1 and p2 in p.

val joinSub : 'c1 Factory.t -> 'c2 Factory.t -> Var.t -> 'c1 t -> 'c2 t -> 'c1 t * 'c2 t

val minkowski : 'c1 Factory.t -> 'c2 Factory.t -> 'c1 t -> 'c2 t -> 'c1 t * 'c2 t

minkowski p1 p2 computes the minkowski sum p of p1 and p2.

val widen : 'c Factory.t -> 'c t -> 'c t -> 'c t

Widening operator. The result includes the two operands, although no certificate is created. Note that widen p1 p2 relies on p1 being included in p2.

val incl : 'c1 Factory.t -> 'c1 t -> 'c2 t -> 'c1 rel_t

incl factory p1 p2 checks the (geometrical) inclusion of p1 in p2.

val inclSub : 'c1 Factory.t -> Var.t -> 'c1 t -> 'c2 t -> 'c1 rel_t

val itvize : 'c Factory.t -> 'c t -> Vec.t -> itvT * 'c option * 'c option

let (itv, lProof, uProof) = itvize p v computes the bounds imposed on the linear form v by the constraints of the polyhedron p. For each bound (upper and lower), an inclusion certificate is returned if it is finite, allowing the prove that the returned interval is indeed correct. lProof and uProof are the certificates for the lower and upper bounds, respectively.

val itvizeSub : 'c Factory.t -> Var.t -> 'c t -> Vec.t -> itvT * 'c option * 'c option

val getUpperBound : 'c Factory.t -> 'c t -> Vec.t -> bndT * 'c option

getUpperBound p v computes the upper bound on the linear form v implied by polyhedron p. If the bound is finite, a certificate is provided.

val getUpperBoundSub : 'c Factory.t -> Var.t -> 'c t -> Vec.t -> bndT * 'c option

val getLowerBound : 'c Factory.t -> 'c t -> Vec.t -> bndT * 'c option

getLowerBound p v computes the lower bound on the linear form v implied by polyhedron p. If the bound is finite, a certificate is provided.

val getLowerBoundSub : 'c Factory.t -> Var.t -> 'c t -> Vec.t -> bndT * 'c option

val rename : 'c Factory.t -> Var.t -> Var.t -> 'c t -> 'c t

rename x y p renames variable x to variable y in polyhedron p. y must be fresh in p.

val invChk : 'c Factory.t -> 'c t -> bool * string

This function checks that the given polyhedron satisfies the following invariants. It returns (true, "") if it does and (false, reason) if not, where reason lists the checks that have failed.

there is no trivial constraint

the polyhedron is not empty

there is no redundant constraint

invariants on the set of equalities

a defined variable appears with unit coefficient

a defined variable does not appear in the equalities before it

invariants on the set of inequalities

none of the variables defined by an equality appears in an inequality

val to_string : (Var.t -> string) -> 'c t -> string

Pretty-printer for a polyhedron.

val to_string_ext : 'c Factory.t -> (Var.t -> string) -> 'c t -> string

val to_string_itv : (Var.t -> string) -> Vec.t -> itvT -> string

val to_string_raw : 'c t -> string

val to_string_ext_raw : 'c Factory.t -> 'c t -> string

val to_string_itv_raw : Vec.t -> itvT -> string

val plot : 'c t -> string

For Sage plot.

val plot_opt : 'c t option -> string

val varSet : 'c t -> Vpl.Var.Set.t

val get_cons : 'c t -> 'c Cons.t list

val get_cstr : 'c t -> Cs.t list

val get_cstr_ineqs : 'c t -> Cs.t list

val equal : 'c1 Factory.t -> 'c2 Factory.t -> 'c1 t -> 'c2 t -> bool

val to_unit : 'c t -> unit t

val get_regions_from_point : 'c Factory.t -> 'c t -> Vec.t -> 'c t list

Returns the partition into regions of the given polyhedron, according to the given normalization point. Certificates are lost during the process: frontiers of regions have no certificate.

val get_regions : 'c Factory.t -> 'c t -> 'c t list

Returns the partition into regions of the given polyhedron. The normalization point is chosen using Opt.get_asg. Certificates are lost during the process: frontiers of regions have no certificate.

val size : 'c t -> Scalar.Rat.t option

Returns an estimation of the size of the polyhedron.

val split_in_half : 'c Factory.t -> 'c t -> Cs.t option

Returns the constraint on which the polyhedron must be split, if there is one. This constraint splits the polyhedron in two halfs along the longuest axis-aligned interval.

val set_point : Vec.t -> 'c t -> 'c t

set_point point p returns the polyhedron p with its interior point set to p.

raises Invalid_argument

if point is not inside p's interior.

val satisfy : 'c t -> Vec.t -> bool

* Returns true if the given point satisfies the constraints of the polyhedron.

val spawn : 'c Factory.t -> 'c t -> Vec.t

* Randomly generates a point in the polyhedron.

val assume_back : 'c Factory.t -> 'c t -> 'c Cons.t -> 'c t option