Module Vpl.Poly

exception Div_by_non_constant

Exception raised when trying to divide a polynomial by a non-constant value

module type Type = Poly_type.Type

Interface of polynomials.

module Make : functor (Vec : Vector.Type) -> Type with module Vec = Vec

Functor to build a polynomial type from a type of vectors. The type of vector coefficients is used for polynomial coefficients. The type of vectors is used to build polynomials from constraints.