Module Type.Vec

module Coeff : Scalar.Type

A type of scalars.

type t = Coeff.t Rtree.t

Type of vectors associating variables to coefficients.

val name : string

Name of the module.

val nil : t

A vector with coefficients all set to zero.

val to_string : (Var.t -> string) -> t -> string

Conversion into string.

parameter var_to_string

a pretty-printer for variables.

val mk : (Coeff.t * Var.t) list -> t

Builds a vector from a list of pairs coefficient * variable.

val toList : t -> (Var.t * Coeff.t) list

Retrieve coefficients and variables of a vector.

val get : t -> Var.t -> Coeff.t

get vec v returns the coefficient of variable v in vec. Returns zero if v is not present in vec.

val set : t -> Var.t -> Coeff.t -> t

set vec v n sets the coefficient of variable v to n in vector vec.

val getVars : t list -> Vpl.Var.Set.t

getvars l returns the set of variables that appear in the list of vectors l.

val eval : t -> t -> Coeff.t

Evaluates a vector on a point.

parameter vec

vector to evaluate

parameter point

evaluation point

val neg : t -> t

neg vec multiplies each coefficient in vec by -1.

val add : t -> t -> t

Adding two vectors.

val sub : t -> t -> t

Substracting two vectors.

val mul_t : t -> t -> t

Computes the multiplication componentwise.

val mulc : Coeff.t -> t -> t

Multiplication by a scalar.

val divc : t -> Coeff.t -> t

Division by a scalar.

val mulr : Scalar.Rat.t -> t -> t

Multiplication by a Rational.

val divr : t -> Scalar.Rat.t -> t

Division by a Rational.

val middle : t -> t -> t

middle x y returns the middle point of the segment xy.

val cmp : t -> t -> int

Syntaxic comparison

val equal : t -> t -> bool

Syntaxic equality

val dot_product : t -> t -> Coeff.t

Dot product of two vectors.

val dot_productr : Scalar.Rat.t Rtree.t -> t -> Coeff.t

Same as dot_product, where the first vector has rational coefficients.

val isomorph : t -> t -> Coeff.t option

isomorph v1 v2 returns Some r if v1 is equal to mult r v2. If there is no such r, None is returned.

val elim : Var.t -> t -> t -> t

elim x v1 v2 eliminates the x from v2 using v1. If v2 has coefficient zero for variable x, it is returned untouched. Otherwise, if variable x has coefficient zero in v1, Invalid_argument "Vec.elim" is raised.

val ofSymbolic : Scalar.Symbolic.t -> Coeff.t

Conversion from a vector with symbolic error.

raises Failure

if called with integer coefficients.

val toRat : t -> Scalar.Rat.t Rtree.t

Conversion into a rational vector.

val ofRat : Scalar.Rat.t Rtree.t -> t

Conversion from a rational vector.

val project : Var.t list -> t -> t

Eliminates a list of variables from a vector, by setting their value to 0.

parameter vars

the variables to eliminate

parameter vec

the vector

val rename : Var.t -> Var.t -> t -> t

Renames a variable.

parameter fromX

the variable to rename

parameter toY

the new variable name

parameter vec

the vector

val rename_f : (Var.t -> Var.t) -> t -> t

Renames all variables of a vector according to a renaming function

parameter f

the renaming function

parameter vec

the vector to rename

val gcd : t -> Q.t

returns

the gcd of the coefficients in the vector.