Module Vpl__CoqPr

Pretty-printing support for Coq standard library types. The Raw version of each function produces a string which describes the structure of its argument in terms of constructors. The primed version produces a Coq string.

val charListTr : char list -> string

val posPr : Vpl.BinNums.positive -> string

val posPrRaw : Vpl.BinNums.positive -> string

val posPr' : Vpl.BinNums.positive -> char list

val posPrRaw' : Vpl.BinNums.positive -> char list

val zPr : Vpl.BinNums.coq_Z -> string

val zPrRaw : Vpl.BinNums.coq_Z -> string

val zPr' : Vpl.BinNums.coq_Z -> char list

val zPrRaw' : Vpl.BinNums.coq_Z -> char list

val charListTr : char list -> string

val stringTr : string -> char list

val posTr : Vpl.BinNums.positive -> Z.t

val zTr : Vpl.BinNums.coq_Z -> Z.t

val exprPr : Vpl.Ring_polynom_AddOnQ.coq_PExpr -> string

val exprPr' : Vpl.Ring_polynom_AddOnQ.coq_PExpr -> char list