Module Vpl__Flags

This module allows to tune the algorithms used in the polyhedral operators.

type scalar =

| Rat

| Symbolic

| Float

Choice of the point type.

Rat: Rational points

Symbolic: Rational points with symbolic error

Float: Floating points

val scalar_to_string : scalar -> string

val scalar_to_xml : scalar -> string

type proj_method =

| FM

| Proj_PLP of scalar

| PHeuristic

Choice of the projection method.

FM: classical way that uses the Fourier-Motzkin algorithm

Proj_PLP: method based on Parametric Linear Programming

type join_method =

| Baryc

| Join_PLP of scalar

| JHeuristic

| Join_fromRegions

Choice of the join method.

Baryc: algorithm based on convex combinations + projection

Join_PLP: method based on Parametric Linear Programming

val proj : proj_method Stdlib.ref

Default choice for projection.

val proj_to_string : unit -> string

val proj_to_xml : unit -> string

val join : join_method Stdlib.ref

Default choice for join.

val join_to_string : unit -> string

val join_to_xml : unit -> string

val handelman_loop : bool Stdlib.ref

If set to true, the Handelman linearization will loop, meaning that it will iterate on the last result obtained. The linearization of polynomial g on the starting polyhedron P will give H_0 = linearize(P,g) H_i = linearize(H_(i-1), g)

val handelman_timeout : int option Stdlib.ref

Sets a timeout to the Handelman linearization. Combined to handelman_loop, the linearization returns the last result found.

val handelman_normalize : Vpl.Scalar.Rat.t option Stdlib.ref

Sets a normalization constant. If None, there will be no normalization.

val distributed_plp : int option Stdlib.ref

If set to Some n, the parametric simplex will be distributed over n processes.

val row_echelon_equalities : bool Stdlib.ref

If set to true, the system of equalities of a polyhedron will be in row echelon form.

type plp_method =

| Adj_Raytracing

| Adj_Raytracing_min

| Greedy

val plp : plp_method Stdlib.ref

val plp_to_string : unit -> string

val plp_to_xml : unit -> string

val log_trace : bool Stdlib.ref

Enabling of log traces.

val sum_lambda_1 : bool Stdlib.ref

If true, put sum lambda = 1 as normalization constraint in the projection via PLP.

type lin_method =

| Intervalization

| Handelman

| Both

val lin : lin_method Stdlib.ref

val smart_proj_incl : bool Stdlib.ref

If set to true, the proj_incl operator will use the smart PLP operator. Otherwise, it computes the projection, then tests the inclusion.