Module Vpl__HLP.Bound

val getWitness : Vpl.Var.t list -> Vpl.HOtypes.LPMaps.t -> Vpl.Var.t -> Vpl.Splx.t -> Vpl.Splx.witness_t option

get vars typ v sx returns a witness allowing to compute the tightest (upper or lower depending on typ bound of v in the polyhedron represented by sx. It returns None if such bound does not exists (i.e. is the polyhedron is unbounded in that direction).

val getBoundIndex : Vpl.Splx.witness_t -> Vpl.HOtypes.Poly.t list -> Vpl.Hi.boundIndex

getBoundIndex wit ph returns the type Hi.boundIndex corresponding to wit, which depends on the order of the constraints in ph.

val get : Vpl.Var.t list -> Vpl.HOtypes.LPMaps.t -> Vpl.Var.t -> Vpl.Splx.t -> Vpl.HOtypes.Poly.t list -> Vpl.Hi.boundIndex option

val updateMaps : Vpl.Var.t list -> (Vpl.Var.t * Vpl.HOtypes.LPMaps.t) list -> Vpl.Splx.t -> Vpl.HOtypes.LPMaps.mapDetBound -> Vpl.HOtypes.LPMaps.mapBound -> Vpl.HOtypes.Poly.t list -> Vpl.HOtypes.LPMaps.mapDetBound * Vpl.HOtypes.LPMaps.mapBound * Vpl.Hi.boundIndex list option