Module Vpl__HLP.Build

val build_equalities : Vpl.Var.t list -> bool -> Vpl.Var.t -> Vpl.HOtypes.CP.t list

There is only one equality in the lp, used to produce a product with the good sign. It is done by counting the number of negative bounds that we have chosen, which is the number of nonnull lpvars_i The equality produced is sum_i y_i = k + (1 or 0 depending on the sign of the monomial to cancel)

val build_inequalities : Vpl.Var.t list -> Vpl.HOtypes.LPMaps.mapDetBound -> Vpl.HOtypes.CP.t list

These inequalities encode that lpvars_i = 1 if the upper bound if x_i is chosen, 0 otherwise.

val build_obj : Vpl.Var.t list -> Vpl.HOtypes.LPMaps.mapDetBound -> Vpl.HOtypes.Poly.Vec.t

The objective function minimizes the use of unknown bounds.

val positivity_cstrs : Vpl.Var.t list -> Vpl.HOtypes.CP.t list

Builds the positivity constraints for each decision variable.

val build : Vpl.Var.t list -> bool -> Vpl.HOtypes.LPMaps.mapDetBound -> Vpl.Splx.t Vpl.Splx.mayUnsatT

Builds the satisfiability problem.