Module Vpl__HOtypes.LPMaps

This module defines the maps used in HLP. They gather information about variable bounds.

type t =

| Upper

| Lower

type mapDetBound = (bool option * bool option) MapV.t

It associates a variable v to a pair of option booleans (b1,b2). b1 = Some true (resp. b2 = Some true) means that v has a lower (resp. upper) bound. b1 = Some false (resp. b2 = Some false) means that v has no lower (resp. upper) bound. b1 = None (resp. b2 = None) means that we don't have information on v's lower (resp. upper) bound.

type mapBound = (Vpl.Hi.boundIndex option * Vpl.Hi.boundIndex option) MapV.t

It associates a variable to a pair of option Hi.boundIndex.

type bounds = {

mapDB : mapDetBound;

mapB : mapBound;

}

val init : Poly.t list -> Vpl.Var.t list -> bounds

init pl vl initializes maps with bounds that appear syntactically in pl.

val hasSup : Vpl.Var.t -> mapDetBound -> Q.t

hasSup v mapDB returns 0 if v has no upper bound in mapDB, 1 otherwise.

val hasInf : Vpl.Var.t -> mapDetBound -> Q.t

hasInf v mapDB returns 0 if v has no lower bound in mapDB, 1 otherwise.

val detSup : Vpl.Var.t -> mapDetBound -> Q.t

detSup v mapDB returns 1 if v has an upper bound in mapDB, 0 otherwise.

val detInf : Vpl.Var.t -> mapDetBound -> Q.t

detInf v mapDB returns 1 if v has a lower bound in mapDB, 0 otherwise.

val mapDB_to_string : mapDetBound -> string

val mapB_to_string : mapBound -> string

val updateMapDB : mapDetBound -> Vpl.Var.t -> bool -> t -> mapDetBound

updateMapDB mapDB v value bound binds in mapDB the upper bound of variable v to value if upper = true. Otherwise, it binds the lower bound of variable v to value.

val updateMapB : mapBound -> Vpl.Var.t -> Vpl.Hi.boundIndex -> t -> mapBound

updateMapB mapB v value bound binds in mapB the upper bound of variable v to value if upper = true. Otherwise, it binds the lower bound of variable v to value.