Module Vpl__HOtypes.MapPolyHi

Module associating in a map a list of Hi.t to a polynomial. It is used to remember what products were used to cancel a given polynomial.

type t = Vpl.Hi.t list MapP.t

Map from polynomials to the list of Handelman products that cancel it.

val to_string : t -> string

Conversion into string.

val memMonom : Poly.MonomialBasis.t -> t -> bool

returns

true if map contains a binding to a polynomial that contains monomial m.

parameter m

a monomial

parameter map

a cancellation map

val memMonomSigned : Poly.Monomial.t -> t -> bool

returns

true if map contains a binding to a polynomial that contains monomial m with a coefficient of the same sign.

val merge : t -> t -> t

Merges two maps.