Module Vpl__Hi

This module defines the type of Handelman products. More precisely, tThey are Schweighofer products.

module Debug : Vpl.DebugTypes.Type

module CP = Vpl.CstrPoly

module Poly = CP.Poly

type cIndex = Vpl.Index.Int.t

Type of Handelman indices. For instance, cIndex (1, 0, 3) represents C_1^1 * C_2^0 * C_3^3.

type varIndex = Vpl.Index.Int.t

Type of indices that represent a product of variables. For instance, varIndex (1, 0, 3) represents x_1^1 * x_2^0 * x_3^3.

type boundIndex = Vpl.Index.Rat.t

Type defining the Farkas decomposition of a constraint. For instance, boundIndex (1, 0, 3) represents 1*C_1 + 0*C_2 + 3*C_3.

type t =

| Ci of cIndex

| VarBounds of varIndex * boundIndex list

| VarCi of varIndex * cIndex

Defines the type of Schweighofer products as either :

A Handelman product using type cIndex

A product of variables bounds (each of them is represented by its Farkas decomposition using type boundIndex) multiplied by a monomial of type varIndex

a Handelman product (type cIndex) multiplied by a monomial of type varIndex

val to_string : t -> string

Conversion into string.

val eq : t -> t -> bool

Equality test.

val is_linear : t -> bool

Is a Handelman product linear.

val computeVarIndex : Vpl.Index.Int.t -> Vpl.Var.t list -> Poly.t

Computes the polynomial corresponding to a variable index.

parameter id

the variable index

parameter the

list of variables

val computeBoundIndex : Vpl.Index.Rat.t -> Poly.t list -> Poly.t

Computes the given Farkas combination.

parameter the

Farkas coefficients

parameter the

constraints

val computeBoundIndexList : Vpl.Index.Rat.t list -> Poly.t list -> Poly.t

Computes a product of Farkas combinations.

parameter the

list of Farkas coefficients

parameter the

constraints

module Cert : sig ... end

Module for building certificates out of Handelman products.