Module Vpl__IOtypes

Types and conversions for intervalization oracle.

type mode = Vpl.DomainInterfaces.mode

module Debug = Vpl.OraclePattern.Debug

module type Domain_T = sig ... end

Type of numerical domains providing linear terms, intervals and polynomials. The intervalization oracle can be run with two domains (in Z when called from Coq, or in Q).

module type Type = sig ... end

Type of domains for intervalization.

module Lift : functor (D : Domain_T) -> sig ... end

Lifts a numerical domain into an intervalization domain.

module DomainZ : sig ... end

Numerical domain for integers.

module DomainQ : sig ... end

Numerical domain for rationals.

module DZ : sig ... end

module DQ : sig ... end