Module Poly.MonomialBasis

MonomialBasis represents the variables of a monomial, i.e. without the coefficient.

type t = (Vpl.Var.t * exp) list

Type of monomial basis.

val null : t

Monomial basis with no variable.

val mk : (Vpl.Var.t * exp) list -> t

Builds a monomial basis.

val add_var : (Vpl.Var.t * exp) -> t -> t

Adds a variable to an already built monomial basis. If the variable is already present, exponents are summed.

val remove_var : Vpl.Var.t -> t -> t

Removes a variable (whatever its exponent) from a monomial basis.

val remove_var_exp : Vpl.Var.t -> int -> t -> t

Removes the given exponent of the variable in the monomial basis.

val to_list : t -> (Vpl.Var.t * exp) list

returns

the monomial basis

val to_string_param : t -> string -> string

Pretty-printer for monomial basis.

returns

the monomial basis where variables (which are integers) are prefixed by s

val to_string : t -> string

Pretty-printer for monomial basis.

returns

the monomial basis where variables (which are integers) are prefixed by "x"

val get_exponent : Vpl.Var.t -> t -> int

returns

the exponent of the given variable.

val compare : t -> t -> int

Lexicographic comparison between two monomial bases.

returns

0 if m1 = m2

returns

-1 if m1 < m2

returns

1 otherwise

val equal : t -> t -> bool

Equality test between two monomial bases.

val sub : t -> t -> t * bool

Removes variables from a monomial basis.

parameter m1

the monomial to substract from

parameter m2

the monomial to substract

returns

true if m2 is a subset of m1 and m1 - m2 makes sense

val rename : t -> Vpl.Var.t -> Vpl.Var.t -> t

Renames a variable.

parameter m

the monomial basis

parameter fromX

the variable to rename

parameter toY

the new variable name

val change_variable : (t -> t option) -> t -> t

Applies the given change of variables on the monomialBasis

val eval : t -> (Vpl.Var.t -> Coeff.t) -> Coeff.t

Evaluates a monomial basis with an evaluation function

parameter m

the monomial basis

parameter f

the evaluation function

val get_vars : t -> Vpl.Var.Set.t

returns

the set of variables of the monomial basis.

val degree : t -> int

returns

the degree of the monomial.

val is_linear : t -> bool

returns

true if the degree of m is at most one. The monomial basis is assumed to be in canonical form.