Module type Vpl__IOtypes.Domain_T

Type of numerical domains providing linear terms, intervals and polynomials. The intervalization oracle can be run with two domains (in Z when called from Coq, or in Q).

module N : Vpl.NumC.NumSig

Type of scalar values.

module BasicTerm : sig ... end

Basic AST Terms, copied from ASTerm.ModalTerm

module LinTerm : sig ... end

Linear terms represented as maps from variables to scalars.

module AffTerm : sig ... end

Affine terms.

module Term : sig ... end

Terms with annotations.

module NoneItv : sig ... end

Not necesarily bounded intervals.

module NA : sig ... end

module NAItv : sig ... end

Type of intervals where bounds are affine terms. Bounds can be unbounded as well.

module Poly : sig ... end

Type fo polynomials.