Module Vpl__Index

This module type defines the type of indexes. An index is a plain vector of coefficients. It is given as a functor that depends on a type of coefficients.

module type Scalar = sig ... end

This module defines the type of Scalar used as coefficients in Indexes

module Make : functor (Coeff : Scalar) -> sig ... end

Functor of Index

module ScalarInt : sig ... end

Type of integer scalars.

module Int : sig ... end

Type of indexes where coefficients are integers.

module Rat : sig ... end

Type of indexes where coefficient type is SxPoly.ScalarQ.t.