Module Vpl__IndexBuild

This module provides efficient decompositions of list of indexes.

module IndexList : sig ... end

module MapI : Stdlib.Map.S with type MapI.key = Vpl.Index.Int.t

module Map : sig ... end

This module associates in a map an Index.Int.t to a IndexList.t. A list il associated to an index i represents a decomposition of i : sum_j il(j) = i.