Module Vpl__IndexBuild.Map

This module associates in a map an Index.Int.t to a IndexList.t. A list il associated to an index i represents a decomposition of i : sum_j il(j) = i.

type t = IndexList.t MapI.t

val to_string : t -> string

val compute : IndexList.t -> t

compute l returns a map associating each index ind of l to a list of indexes whose sum is ind. Indexes that have only one non null coefficient are not added into the map.

raises Invalid_argument

if l is empty.

val compute_list_from_map : IndexList.t -> t -> t

Same as function compute but starts from a given map.

raises Invalid_argument

if the input list is empty.

val compute_from_map : Vpl.Index.Int.t -> t -> IndexList.t * t

Same as function compute_list_from_map but with a single index instead of a list. Returns the corresponding type IndexList.t as well.