Module Vpl__Min.Make

Parameters

VecInput : Vpl.Vector.Type

Vec : Vpl.Vector.Type

CsInput : Vpl.Cstr.Type

LP : Vpl.MinLP.Type

Signature

include Type with module VecInput = VecInput

module VecInput = VecInput

val minimize : VecInput.t -> Vpl.Cstr.Rat.t list -> (Vpl.Cstr.Rat.t * VecInput.t) list

minimize x cstrs removes the redundancies in the list of constraints cstrs.

parameter cstrs

is the list of constraints to minimize. Syntactic redundancies should have been removed previously.

parameter x

is a point that should lie in the interior of the polyhedron defined by cstrs. This function attach to each returned constraint cstr a point that violates only cstr. When possible, this point does not saturate the completentary of cstr.

val minimize_cone : VecInput.t -> Vpl.Cstr.Rat.t list -> (Vpl.Cstr.Rat.t * VecInput.t) list

val name : string

type conversion = {

vec_CsVec : Vec.t -> LP.CsUser.Vec.t;

csVec_Vec : LP.CsUser.Vec.t -> Vec.t;

csCoeff_Coeff : LP.CsUser.Vec.Coeff.t -> Vec.Coeff.t;

csInput_Cs : CsInput.t -> LP.CsUser.t;

vecInput_Vec : VecInput.t -> Vec.t;

vec_VecInput : Vec.t -> VecInput.t;

}

val conv : conversion Stdlib.ref

type direction_type =

| Normal of Vec.t * LP.CsUser.Vec.t
x' = x0 + t * normal

| TwoPoints of Vec.t * Vec.t
x' = x0 + t * (x1 - x0)

type direction = direction_type * LP.CsUser.Vec.Coeff.t

module Sort : sig ... end