Module Type.CsUser

module Vec : Vpl.Vector.Type

A module of vectors.

module Coeff : Vpl.Scalar.Type

The coefficients of Vec.

type t = {

typ : Vpl.Cstr_type.cmpT;
the comparison operator

v : Vec.t;
the linear term

c : Vec.Coeff.t;
the constant

}

Type of affine constraints, of the form c.v c.typ c.c.

val name : string

Module name.

val top : t

0 = 0

val get_typ : t -> Vpl.Cstr_type.cmpT

Returns the contraint comparator.

val get_v : t -> Vec.t

Returns the constraint linear term.

val get_c : t -> Vec.Coeff.t

Returns the constraint constant term.

val to_string : (Vpl.Var.t -> string) -> t -> string

Pretty-printer for constraints.

parameter var_to_string

a pretty-printer for variables

parameter cstr

the constraint to print

val list_to_string : t list -> string

Pretty-printers for lists of constraints.

val cmpAdd : Vpl.Cstr_type.cmpT -> Vpl.Cstr_type.cmpT -> Vpl.Cstr_type.cmpT

Computes the comparator resulting from the addition of two comparators.

val compl : t -> t

Compute the complement constraint to the given inequality.

raises Invalid_argument

if the supplied linear constraint is an equality

val split : t -> t * t

Splits the given equality (e.g. x = 1) into the two equivalent inequalities (e.g. x <= 1 and -x <= -1).

raises Invalid_argument

if the supplied constraint is not an equality

val mk : Vpl.Cstr_type.cmpT -> (Vec.Coeff.t * Vpl.Var.t) list -> Vec.Coeff.t -> t

Builds a linear constraint

parameter cmp

the comparison operator

parameter l

the list of pairs (coefficient * variable), representing the linear term. Zero coefficients can be left implicit.

parameter cst

the constant term

val mk2 : Vpl.Cstr_type.cmpT -> Vec.t -> Vec.Coeff.t -> t

Builds a linear constraint

parameter cmp

the comparison operator

parameter vec

the linear term

parameter cst

the constant term

val eq : (Vec.Coeff.t * Vpl.Var.t) list -> Vec.Coeff.t -> t

Builds an equality.

parameter l

the list of pairs (coefficient * variable), representing the linear term. Zero coefficients can be left implicit.

parameter cst

the constant term

val le : (Vec.Coeff.t * Vpl.Var.t) list -> Vec.Coeff.t -> t

Builds a large inequality.

parameter l

the list of pairs (coefficient * variable), representing the linear term. Zero coefficients can be left implicit.

parameter cst

the constant term

val lt : (Vec.Coeff.t * Vpl.Var.t) list -> Vec.Coeff.t -> t

Builds a strict inequality.

parameter l

the list of pairs (coefficient * variable), representing the linear term. Zero coefficients can be left implicit.

parameter cst

the constant term

val mkInt : Vpl.Cstr_type.cmpT -> (Vec.Coeff.t * Vpl.Var.t) list -> Vec.Coeff.t -> t

Same as mk but transforms coefs < cst into coefs <= (cst - 1).

parameter cmp

the comparison operator

parameter l

the list of pairs (coefficient * variable), representing the linear term. Zero coefficients can be left implicit.

parameter cst

the constant term

val equalSyn : t -> t -> bool

Checks the syntactical equality between two constraints. x = 1 and x = 1 are syntactically equal, but not 2 * x = 2 and x = 1.

val inclSyn : t -> t -> bool

Check the syntactical inclusion of one constraint in another. The linear term of both constraints must be syntactically equal. x = 0 is included in x < 1, but not is 2 * x < 2.

val equal : t -> t -> bool

Checks if two constraints describe the same space.

val incl : t -> t -> bool

Checks if the subspace described by c1 is included in that of c2.

val cmp : t -> t -> int

Comparison between to constraints.

val add : t -> t -> t

Adds two constraints.

val mulc : Vec.Coeff.t -> t -> t

Multiplies a constraint by a constant.

raises BadMult

if a negative constant is provided and the constraint is an inequality

val mulc_no_exc : Vec.Coeff.t -> t -> t

Same as mulc but works for negative constants as well.

type prop_t =

| Trivial
0 = 0 or 0 < 1 are Trivial

| Contrad
trivialy contradictory: e.g. 0 = 1 or 0 < -1

| Nothing
neither trivial nor trivialy contradictory

The type of properties the can be told by examining a constraint on its own.

val tellProp : t -> prop_t

Tells what can be infered on a constraint on its own. See prop_t

val getVars : t list -> Vpl.Var.Set.t

Builds the set of variables which appear with non-zero coefficients in the constraints.

val getCoefsFor : Vpl.Var.t option -> t list -> Vec.Coeff.t list

getCoefsFor mx l gathers the "coefficient" of mx in the constraints in l. If mx is None then the constants are gathered. If mx is Some x, the coefficients of variable x are gathered. The order of the numbers in the returned list matches that of the constraints in l.

val satisfy : Vec.t -> t -> bool

satisfy vec cstr returns true if point vec satisfies constraint cstr.

val saturate : Vec.t -> t -> bool

saturate vec cstr returns true if vec satisfies constraint cstr where cstr.typ has been replaced by Cstr_type.cmpT.Eq.

val eval : t -> Vec.t -> Vec.Coeff.t

Evaluates a constraint on a point. Computes a.point - b, where c : ax cmp b.

val rename : Vpl.Var.t -> Vpl.Var.t -> t -> t

Renames a variable.

parameter fromX

the variable to rename

parameter toY

the new variable name

parameter cstr

the constraint

val rename_f : (Vpl.Var.t -> Vpl.Var.t) -> t -> t

Renames all variables of a constraint according to a renaming function

parameter f

the renaming function

parameter vec

the vector to rename

val change_variable : Vpl.Var.t -> Vec.t -> Vec.Coeff.t -> t -> t

change_variable x lin c cstr applies the change of variable x = lin + c in cstr.

val get_saturating_point : t -> Vec.t

returns

a point that saturates the hyperplane defined by the given constraint.

val distance_point_cstr : Vec.t -> t -> Vec.Coeff.t

Returns the euclidian distance between the given point and the line defined by the constraint.

val plot : Vpl.Var.t list -> t -> string

/

val list_plot : t list -> string

/