Module Vpl__Misc

This module defines several generic functions to handle lists or strings

val sublist : 'a list -> int -> int -> 'a list

sublist l i j returns the sublist of l starting at index i and ending at index j-1

val index : string -> char -> int

index s c returns the index of the character c in the string s if it exists, -1 otherwise

val substring : string -> int -> string

substring s i returns the substring of s starting at index i

val findi : ('a -> bool) -> 'a list -> int

findi p l returns the index of the first element in l that satisfies predicate p.

raises Not_found

if no such element exists

val array_findi : (int -> 'a -> bool) -> 'a array -> int

val find_res : ('a -> bool * 'a) -> 'a list -> 'a

val popi : 'a list -> int -> 'a list

popi l i returns the list l where the ith element has been removed. If i < 0 or i > len(l), it returns l

val pop : ('a -> 'a -> bool) -> 'a list -> 'a -> 'a list

pop eq l x returns the list l where the first element equal to x (w.r.t equality function eq) is removed

val popAll : ('a -> 'a -> bool) -> 'a list -> 'a -> 'a list

pop eq l x returns the list l where all elements equal to x (w.r.t equality function eq) are removed

val list_to_string : ('a -> string) -> 'a list -> string -> string

list_to_string to_string l sep returns the string to_string l(0) ^ sep ^ ... ^ sep ^ to_string l(len(l)-1)

val array_to_string : ('a -> string) -> 'a array -> string -> string

array_to_string to_string l sep returns the string to_string l(0) ^ sep ^ ... ^ sep ^ to_string l(len(l)-1)

val list_eq : ('a -> 'a list -> bool) -> 'a list -> 'a list -> bool

list_eq contains l1 l2 returns true if l1 and l2 contain the same elements. contains x l returns true if x belongs to l, false otherwise

val list_eq2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val range : int -> int -> int list

range a b returns the list of consecutive integers from a to b-1

val max : ('a -> 'a -> int) -> 'a list -> 'a

max cmp l returns the maximum element of l w.r.t the comparison function cmp

raises Invalid_argument

if the input list is empty

val maxi : ('a -> 'a -> int) -> 'a list -> int

maxi cmp l returns the index of the maximum element of l w.r.t the comparison function cmp

raises Invalid_argument

if the input list is empty

val min : ('a -> 'a -> int) -> 'a list -> 'a

min cmp l returns the minimum element of l w.r.t the comparison function cmp

raises Invalid_argument

if the input list is empty

val rem_dupl : ('a -> 'a -> bool) -> 'a list -> 'a list

rem_dupl eq l removes the multiple occurences of elements in l, w.r.t equality function eq

val intersection : ('a -> 'a -> bool) -> 'a list -> 'a list -> 'a list

intersection eq l1 l2 returns the list of elements that are both in l1 and l2, w.r.t equality function eq

val is_sublist : 'a list -> 'a list -> bool

sublist l1 l2 returns true if l1 is a sublist of l2, false otherwise

val map2i : (int -> 'a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val filter_i : ('a -> bool) -> 'a list -> int -> 'a list * 'a list

filter_i f l i returns the i first elements of l that are true through f

val string_repeat : string -> int -> string

string_repeat s i returns a string composed of i times s

val add_tab : int -> string -> string

add_tab i s adds i tabulations after each '\n' in s.

val fold_left_i : (int -> 'a -> 'b -> 'a) -> 'a -> 'b list -> 'a

val fold_right_i : (int -> 'a -> 'b -> 'b) -> 'a list -> 'b -> 'b

val array_fold_left_i : (int -> 'a -> 'b -> 'a) -> 'a -> 'b array -> 'a

val string_equal : string -> string -> bool

val init_list : int -> (int -> 'a) -> 'a list

Builds a list from a length and an initialization function.

parameter len

the list length

parameter f

an initialization function: index i is initialized as f i

parameter raise

Invalid_argument if len < 1

val subtract : 'a list -> 'a list -> 'a list