Module Vpl__OraclePattern

Interface of linearization Oracles.

module Debug : sig ... end

module type OracleType = sig ... end

Interface of linearization Oracles. An Oracle has a modular design: strategies are encoded within Prayers, that are ordered in a list (prayers). Starting from a polynomial to treat, each prayer is tried one after the other. If it can be applied, the prayer updates the polynomial until it is 0.

module Make : functor (Oracle : OracleType) -> sig ... end

Builds a runnable oracle from an OracleType.