Module Vpl__OraclePattern.Make

Builds a runnable oracle from an OracleType.

Parameters

Oracle : OracleType

Signature

include Oracle

module P = Oracle.P

Module of polynomials

type prophecy

Prophecy made by the Oracle.

module type Prayer = sig ... end

Prayer interface.

val prayers : (module Prayer) list

List of prayers. There are applied in this order.

val recursive_oracle : (P.t -> prophecy -> (module Prayer) list -> prophecy) Stdlib.ref

Pointer to the oracle, to allow recursive calls within the oracle itself.

val my_prayers : (module Prayer) list Stdlib.ref

List of prayers used in the oracle. It is initialized with prayers.

val remove_prayer : (module Prayer) -> unit

val make_prophecy : P.t -> prophecy -> (module Prayer) list -> prophecy

Runs the oracle until the polynomial is 0. Prayers are tried in the order given by prayers. If a prayer is applicable, the oracle starts back with the initial list of prayers.

parameter p

the polynomial to handle

parameter pr

the current prophecy

parameter prayers

the current prayers list

returns

the new prophecy

val run : prophecy -> P.t -> prophecy

Runs the oracle.

parameter init_prophecy

the initial prophecy

parameter p

the polynomial to handle

returns

the prophecy