Module type Vpl__OraclePattern.OracleType

Interface of linearization Oracles. An Oracle has a modular design: strategies are encoded within Prayers, that are ordered in a list (prayers). Starting from a polynomial to treat, each prayer is tried one after the other. If it can be applied, the prayer updates the polynomial until it is 0.

module P : Vpl.Poly.Type

Module of polynomials

type prophecy

Prophecy made by the Oracle.

module type Prayer = sig ... end

Prayer interface.

val prayers : (module Prayer) list

List of prayers. There are applied in this order.

val recursive_oracle : (P.t -> prophecy -> (module Prayer) list -> prophecy) Stdlib.ref

Pointer to the oracle, to allow recursive calls within the oracle itself.