Module Vpl__PLPCore

module Cs = Vpl.Cstr.Rat

module Debug : sig ... end

module Profile : sig ... end

module Vec = Cs.Vec

module Minimization : sig ... end

module Explore = Vpl.PSplxExec.Explore

module Coeff = Vec.Coeff

module MapV : sig ... end

type region_t =

| Cone

| NCone

module Boundary : sig ... end

val strict_comp : Cs.t -> Cs.t

val cstr_eq : Cs.t -> Cs.t

val is_trivial : Cs.t -> bool

module ExplorationPoint : sig ... end

module Region : sig ... end

type 'c mapRegs_t = 'c Region.t MapV.t

type 'c t = {

regs : 'c mapRegs_t;

todo : ExplorationPoint.t list;

}

val empty : 'c t

val to_string : 'c t -> string

module Check : sig ... end

Module that checks results on the fly. It is based on Debug.Check, thus these verifications cost almost nothing if checking is disabled.

module Adjacency : sig ... end

module Next_Point : sig ... end

Several methods to choose the next point to explore.

val get_next_point : region_t -> 'a t -> ExplorationPoint.t option * 'a t

val reg_id : int Stdlib.ref

val get_fresh_id : unit -> int

Returns a fresh id for a new region.

module Exec : sig ... end

Module that runs one exploration

val extract_points : 'c Region.t -> int -> ExplorationPoint.t list

type 'c config = {

add_region : 'c Region.t option -> 'c Region.t -> ExplorationPoint.t -> 'c t -> 'c t;

reg_t : region_t;

points : ExplorationPoint.t list;

regions : 'c Region.t list;

}

module Add_Region : sig ... end

val init_sx : 'c Vpl.Factory.t -> 'c Vpl.PSplx.t -> Vec.t -> bool

Leads to a feasible parametric simplex tableau.

val find_new_point : Vec.t -> Vec.t -> Vec.t

This function is used when an execution (ran on point p) led to a region with empty interior. It returns the middle between the initial region's point and p.

val exec : 'c config -> 'c Vpl.PSplx.t -> 'c t -> 'c t

val result_to_string : ('c Region.t * 'c Vpl.Cons.t) list option -> string

val get_results : 'c Vpl.Factory.t -> 'c t -> ('c Region.t * 'c Vpl.Cons.t) list

Careful: if a region has no simplex tableau (because it was given in input), it does not appear in the result.

val init_regions : 'c config -> 'c Vpl.PSplx.t -> 'c mapRegs_t

val choose_init_point : 'c Vpl.PSplx.t -> 'c config -> Vec.t * 'c config

val init_and_exec : 'c Vpl.Factory.t -> 'c config -> 'c Vpl.PSplx.t -> 'c t option

val run : 'c Vpl.Factory.t -> 'c config -> 'c Vpl.PSplx.t -> ('c Region.t * 'c Vpl.Cons.t) list option