Module Vpl__PLPCore.Add_Region

val should_explore_again : 'c Region.t -> Cs.t -> 'c Region.t -> bool

should_explore_again reg1 cstr reg2 checks whether the point that has been explored should be explored again.

parameter reg1

is the provenance region.

parameter cstr

is constraint that has been crossed in the previous region.

parameter reg2

is the new discovered region.

val region_already_known : 'c Region.t -> 'c t -> bool

region_already_known reg plp checks whether reg is already present in plp.

val has_interior : 'c Region.t -> bool

Return a witness if the region has no interior, None otherwise.

val remove_point : ExplorationPoint.t -> ExplorationPoint.t list -> ExplorationPoint.t list

val do_not_add : 'c Region.t -> ExplorationPoint.t -> 'c t -> 'c t

val add : 'c config -> 'c Region.t option -> 'c Region.t -> ExplorationPoint.t -> 'c t -> 'c t