Module Vpl__PLPCore.Coeff

type t = Q.t

Type of scalar value

val name : string

Name of the scalar type.

val z : t

Zero value.

val u : t

Value 1.

val negU : t

Value -1.

val delta : t

Small value used for conversions from Scalar.symbolic.

val cmp : t -> t -> int

cmp x y returns 0 if x = y, a negative integer if x < y, and a positive one otherwise.

val le : t -> t -> bool

le x y returns true if x <= y, false otherwise.

val lt : t -> t -> bool

lt x y returns true if x < y, false otherwise.

val cmpz : t -> int

cmpz n is equivalent to cmp z n.

val equal : t -> t -> bool

equal x y returns true if x = y, false otherwise.

val equalApprox : t -> t -> bool

equalApprox x y returns true if abs (x - y) <= delta.

val isZ : t -> bool

isZ n returns true is n is zero (z) and false otherwise.

val neg : t -> t

Compute the opposite of a number.

val inv : t -> t

Compute the inverse of a number.

val abs : t -> t

Compute the absolute value.

val add : t -> t -> t

Addition of two constants.

val sub : t -> t -> t

Substraction of a constant by another.

val mul : t -> t -> t

Multiplication of two constants.

val pow : t -> int -> t

Power of a constant.

val div : t -> t -> t

div num den

val mulr : Q.t -> t -> t

Multiplication by a rational.

val divr : t -> Q.t -> t

Division by a rational.

val mk : int -> int -> t

mk ~den:i1 i2 builds an element of type t and of value i2/i1.

val of_int : int -> t

Builds an element of type t from an integer.

val to_int : t -> Vpl.Scalar_type.roundT -> int option

to_int x Up returns the smallest i such that i >= x. to_int x Down returns the greatest i such that i <= x. Returns None if x does not fit within a machine integer.

val of_float : float -> t

Conversion from float.

val to_float : t -> float

Conversion to float.

val toZ : t -> Z.t * Z.t

Conversion into integer.

val ofZ : Z.t -> Z.t -> t

Conversion from integer.

val ofQ : Q.t -> t

Conversion from rational.

val toQ : t -> Q.t

Conversion into rational.

val to_string : t -> string

Conversion into string.

val of_string : string -> t

Conversion from string.

val well_formed : t -> bool

well_formed x returns true if x is not +/- infinity

val well_formed_nonnull : t -> bool

well_formed_nonnull x returns true if x is neither 0 nor +/- infinity