Module Vpl__PLPCore.Exec

Module that runs one exploration

val correct_point : region_t -> Cs.t list -> Vec.t option

correct_point cstrs point returns a point that lies in cstrs's interior. If no such point exists, None is returned. It uses Region.getPointInside.

val get_boundaries : region_t -> Cs.t list -> Vec.t -> Boundary.t list

get_boundaries cstrs point returns the list of boundaries by minimizing cstrs.

parameter point

must be a point in the interior of cstrs.

val exec' : region_t -> 'c Vpl.PSplx.t -> Vec.t -> 'c Region.t option

val exec : region_t -> 'c Vpl.PSplx.t -> Vec.t -> 'c Region.t option

exec strat sx map exp returns the region resulting from the exploration of exp. If this region has empty interior, it returns None.