Module Vpl__PLPCore.Next_Point

Several methods to choose the next point to explore.

module Min : sig ... end

val conv : Min.conversion

module Point_in_Region : sig ... end

module RayTracing : sig ... end

Module that chooses the next point to explore by raytracing.

module RayTracing_min : sig ... end

Other algorithm to find new exploration points

module Greedy : sig ... end