Module Next_Point.RayTracing_min

Other algorithm to find new exploration points

val eval : (int * 'c Region.t) list -> Min.direction_type -> ((int * Cs.t) * Min.direction) list

val debug_evals : ((int * Cs.t) * ('a * Cs.Vec.Coeff.t)) list -> ((int * Cs.t) * ('a * Cs.Vec.Coeff.t)) list

val compute_next_point : Cs.t -> (('a * Cs.t) * Min.direction) list list -> Cs.Vec.Coeff.t option

val eval_to_string : (('a * Cs.t) * Min.direction) list list -> string

exception Adjacent of int

val adjust : region_t -> 'c mapRegs_t -> (int * Boundary.t) -> (int * 'c Region.t) -> ExplorationPoint.t

val apply_adjacency : 'c t -> int -> Boundary.t -> int -> 'c t

val get_next_point_rec : region_t -> 'c t -> 'c t

The PLP state t is modified if an adjacency is found.

val get_next_point : region_t -> 'c t -> ExplorationPoint.t option * 'c t

Returns the next point to explore, or None if there is none. The returned point is the first element of plp.todo.