Module Vpl__PSplx

This module defines the type of parametric simplex tableau. In VPL, the PLP is always use to express a combination of some input linear constraints. This specificity is thus put directly in the simplex tableau. Each column of the tableau is associated with an input constraint (with a certificate). Constraints of the problem are expressed as functions of these constraints.

module Debug : Vpl.DebugTypes.Type

module Cs = Vpl.Cstr.Rat

type decision_variable = int

Each column of the simplex tableau is associated with a decision variable. Therefore, decision variables are identified by their column position in the simplex tableau.

type var_set = int

Decision variables can be given a set of variables, with which they are allowed to pivot. By default, variables are given the set 1.

module VarMap : Stdlib.Map.S with type VarMap.key = decision_variable

type pivotT = (Cs.t * int * Vpl.Tableau.t) -> Cs.t

This type is used to replay pivots when adding new columns. It represents a pivot operation.

parameter cstr

the objective coefficient

parameter i_col

the index of the new column

parameter tab

the simplex tableau on which the pivot must be applied

returns

the new objective coefficient

type 'c t = {

mutable obj : Vpl.Objective.t;
Objective function

mutable tab : Vpl.Tableau.t;
Constraint matrix

mutable get_set : var_set VarMap.t;
map associating decision variables to their set

mutable basis : int list;

mutable sets : var_set list;
Associates a var_set to each row

cstrs : 'c Vpl.Cons.t list;
The constraint associated to each variable.

mutable new_col : (Cs.t -> Vpl.Scalar.Rat.t) array;

mutable pivots : pivotT list;
List of pivots performed on this tableau.

}

Type of simplex tableau. All fields that may change during a pivoting operation are mutable.

val empty : 'c t

Empty simplex tableau.

val copy : 'c t -> 'c t

val copy_in : 'c t -> 'c t -> unit

val to_string : 'c t -> string

Conversion intro string.

val print : 'c t -> unit

Prints the simplex tableau in the standard output.

val nVars : 'c t -> int

returns

the number of decision variables of the problem.

val nParams : 'c t -> int

returns

the number of parameters of the problem.

val nRows : 'c t -> int

returns

the number of rows of the problem.

val getParams : 'c t -> Vpl.Var.Set.t

returns

the set of parameters of the problem.

val objValue : 'c t -> Cs.t

returns

the current value of the objective function, expressed as a constraint over the parameters.

val objValueCert : 'c Vpl.Factory.t -> 'c t -> 'c Vpl.Cons.t

val constant_index : 'c t -> int

val getCurVal : 'c t -> (int * Q.t) list

getCurVal sx returns the current value of the basic variables in sx. The value of the other variables is implicitly zero. Each basic variable is given its value as a pair (column number, value)

val isFeasible : 'c t -> bool

returns

true is the basic solution is feasible, i.e. if all variables have a nonnegative value. The basic solution consists in setting all non-basic variables to zero and setting each basic variable to the right-hand side of the constraint it appears on.

val remCol : int -> 'c t -> 'c t