Module Vpl__PSplxBuild

This modules allows to build a simplex tableau.

val trivial_constraint : 'c Vpl.Factory.t -> 'c Vpl.Cons.t

val update_new_col : int -> (Vpl.PSplx.Cs.t -> Vpl.Scalar.Rat.t) -> 'c Vpl.PSplx.t -> unit

val init : 'c Vpl.Cons.t list -> int -> int list -> Vpl.PSplx.Cs.Vec.t -> 'c Vpl.PSplx.t

Initializes a simplex tableau.

parameter cstrs

the list of constraint associated with each column

parameter n_rows

the number of rows

parameter obj_cstrs

the indices of objective constraints in cstrs

parameter normalization_point

the normalization point

val elim_from : int -> Vpl.Var.t -> int list -> Vpl.PSplx.var_set -> 'c Vpl.PSplx.t -> unit

Adds a constraint to eliminate a variable. If the variable does not appear in any constraint, no row is added.

parameter i_row

the row index

parameter var

the variable to eliminate

parameter i_cstrs

the constraint indices from which the variable is eliminated

parameter sx

the simplex tableau

val var_equal : int -> Vpl.Var.t -> int list -> int list -> Vpl.PSplx.var_set -> 'c Vpl.PSplx.t -> unit

Adds an equality constraint between the coefficients of a variable in two sets of constraints. If the variable does not appear in any constraint, no row is added.

parameter i_row

the row index

parameter var

the variable

parameter i_cstrs1

a set of constraints

parameter i_cstrs2

a second set of constraints

parameter sx

the simplex tableau

val cste_equal : int -> int list -> int list -> Vpl.PSplx.var_set -> 'c Vpl.PSplx.t -> unit

Adds an equality constraint between the constants in two sets of constraints. If the variable does not appear in any constraint, no row is added.

parameter i_row

the row index

parameter i_cstrs1

a set of constraints

parameter i_cstrs2

a second set of constraints

parameter sx

the simplex tableau

val set_var_set : Vpl.PSplx.decision_variable -> Vpl.PSplx.var_set -> 'c Vpl.PSplx.t -> unit

Set the variable set of a row.

val add_col : 'c Vpl.Cons.t -> 'c Vpl.PSplx.t -> 'c Vpl.PSplx.t

Adds a new column to the simplex tableau. The column is added at the end of the tableau. The variable set of the column is necessarily 1.