Module Vpl__PSplxExec

This modules contains functions for executing the simplex algorithm on a parametric sumplex tableau.

module Cs = Vpl.Cstr.Rat

exception Unbounded_problem

Exception raised when a problem is unbounded.

exception Infeasible_problem

Exception raised when a problem is infeasible.

module Explore : sig ... end

module Init : sig ... end

val init_and_push : 'c Vpl.Factory.t -> Cs.Vec.t -> 'c Vpl.PSplx.t -> bool

Initializes the simplex tableau and launch the optimization.

parameter factory

the certified factory

parameter point

the parametric point on which the objective function is instantiated

parameter sx

the simplex tableau

returns

false if the problem is infeasible, true otherwise