Module Vpl__Rtree

Module of radix trees.

type 'n t =

| Nil
Leaf

| Sub of 'n t * 'n * 'n t
Left branch, value, right branch.

val empty : 'n t

Empty map.

val is_empty : 'n t -> bool

Returns true if the map is empty.

val get : 'n -> 'n t -> Vpl.Var.t -> 'n

get z map v returns the value associated to v in map. If there is no such binding, it returns z.

val set : 'n -> 'n t -> Vpl.Var.t -> 'n -> 'n t

set z map var value returns map, with a binding from var to value. If the implemented map is Rtree, value z is placed on each intermediate node.

val mk : 'n -> (Vpl.Var.t * 'n) list -> 'n t

mk z l builds a map from the list of association l. Is uses set that needs a value z for the implementation Rtree.

val map : ('n -> 'm) -> 'n t -> 'm t

map f m applies function f for each binding in m.

val fold : (Vpl.Var.t -> 'a -> 'n -> 'a) -> 'a -> 'n t -> 'a

val fold2 : (Vpl.Var.t -> 'a -> 'n -> 'm -> 'a) -> 'a -> 'n t -> 'm t -> 'a

Same as fold, but with two maps. Function identity is applied when a binding is missing in one map.

val fold2_opt : (Vpl.Var.t -> 'a -> 'n option -> 'm option -> 'a) -> 'a -> 'n t -> 'm t -> 'a

Same as fold2, but the function uses option types.

val find : ('a -> 'b option) -> 'a t -> (Vpl.Var.t * 'b) option

find f rt looks for a node n of rt for which f returns Some b. If such a node is found, the path to n in rt is returned along with the b returned by f. If f returns None for all the nodes of rt, the function find returns None. The order in which find walks through rt is not specified.

val find2 : (Vpl.Var.t -> 'm -> 'n -> 'b option) -> 'm t -> 'n t -> (Vpl.Var.t * 'b) option

Same as find with two maps.

val findPred : ('a -> bool) -> 'a t -> (Vpl.Var.t * 'a) option

Same as find where the given function is a predicate.

val findPred2 : ('a -> 'b -> bool) -> 'a t -> 'b t -> (Vpl.Var.t * 'a * 'b) option

Same as find2 where the given function is a predicate.

val toList : 'a t -> (Vpl.Var.t * 'a) list

val to_string : string -> ('a -> string -> string) -> (Vpl.Var.t -> string) -> 'a t -> string

val mskBuild : ('a -> bool) -> 'a t list -> bool t

mskBuild pred l builds a mask tree where a node is true if pred has returned true for at least one tree in l for that node.

val pathsGet : bool t -> Vpl.Var.Set.t

pathsGet m returns the set of variables which value is true in map m.

val basisBuild : ('a -> bool) -> (Vpl.Var.t -> string) -> 'a t list -> string t

basisBuild isNil pr l builds a string t where a path leads to its associated string (provided by pr) if there is at least a tree in l which associates a non-nil coefficient (determined by isNil) to this path.

val merge : (Vpl.Var.t -> 'a option -> 'b option -> 'c option) -> 'a t -> 'b t -> 'c t

val merge3 : (Vpl.Var.t -> 'a option -> 'b option -> 'c option -> 'res option) -> 'a t -> 'b t -> 'c t -> 'res t

val for_all2 : ('m option -> 'n option -> bool) -> 'm t -> 'n t -> bool

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

equal cmp m1 m2 tests whether the maps m1 and m2 are equal, that is, contain equal keys and associate them with equal data.

parameter cmp

is the equality predicate used to compare the data associated with the keys.