Module Vpl__Splx

This module implements a feasibility-only simplex algorithm. The basis for the algorithm comes from the article "Integrating simplex with DPLL(T)" published by Bruno Dutertre and Leonardo de Moura in 2006 as an SRI International technical report. The interface of this module needs a fair amount of user-friendliness.

Remarks. This module works only with variables of Var.Positive. Its implementation directly depends on the data structure Rtree.

type bnd_t = {

id : int;

scale : Vpl.Scalar.Rat.t;

bv : Vpl.Scalar.Symbolic.t;

}

data type for the description of bound

The id field records which input constraint the bound is the result of. The scale field is used to generate correct certificates in case the input constraint was of the form 2.x <= 2. The bv field is the actual value of the bound of the variable.

module Cs = Vpl.Cstr.Rat

module Vec = Cs.Vec

val get_id : bnd_t -> int

accessor for the id field of a value of type bnd_t

val get_scale : bnd_t -> Vpl.Scalar.Rat.t

accessor for the scale field of a value of type bnd_t

val get_bv : bnd_t -> Vpl.Scalar.Symbolic.t

accessor for the bv field of a value of type bnd_t

val bnd_equal : bnd_t -> bnd_t -> bool

equality test on values of type bnd_t.

val obnd_equal : bnd_t option -> bnd_t option -> bool

equality test on values of type bnd_t option.

type state_t = {

v : Vpl.Scalar.Symbolic.t;

low : bnd_t option;

up : bnd_t option;

}

val get_v : state_t -> Vpl.Scalar.Symbolic.t

val get_low : state_t -> bnd_t option

val get_up : state_t -> bnd_t option

val state_equal : state_t -> state_t -> bool

equality test on values of type state_t.

module Defs : sig ... end

type t = {

mat : Vec.t option Vpl.Rtree.t;

state : state_t Vpl.Rtree.t;

vars : Vpl.Var.Set.t;

defs : Defs.t;

nxt : Vpl.Var.t;

}

The type of a simplex tableau.

val get_mat : t -> Vec.t option Vpl.Rtree.t

val get_state : t -> state_t Vpl.Rtree.t

val getVars : t -> Vpl.Var.Set.t

val get_nxt : t -> Vpl.Var.t

module Witness : sig ... end

type witness_t = Witness.t

type 'a mayUnsatT =

| IsOk of 'a

| IsUnsat of witness_t

val add : t -> (int * Cs.t) -> t mayUnsatT

add s (id, c) adds contraints c, identified by id, to the simplex problem s.

add assumes, but does not check, that id uniquely identifies c. If the addition of c to the simplex problem makes it unsatisfiable, add may or may not return IsUnsat w, where w describes a linear combination of the input constraints yielding a contradiction, otherwise, the updated problem is returned.

In the current implementation, IsUnsat is only returned when specifying incompatible bounds on single variables, e.g. x < 1 and x > 1.

val addAcc : t mayUnsatT -> (int * Cs.t) -> t mayUnsatT

addAcc is a convenience wrapper around add which propagates IsUnsat or performs the addition using add as necessary.

val mk : Vpl.Var.t -> (int * Cs.t) list -> t mayUnsatT

mk lim l builds a new simplex problem, composed of the constraints in l.

Variable lim specifies a frame beyond which fresh variables can be allocated, using Var.next. Note that lim itself may be used as a fresh variable. There is no possibility to change the frame after a simplex data structure has been created using mk: all the constraints ever added to the result of mk should have their variables "smaller" than lim.

mk is implemented in terms of add and has the same behaviour regarding IsUnsat return values.

val check : t -> t mayUnsatT

Look for an assignment of the variables of the problem which satisfies all the constraints. If such an assignment is found, the last simplex tableau is returned. If no such assignment exist, an unsatifiability witness is returned which describes a linear combination of the input constraints yielding a contradiction.

val checkFromAdd : t mayUnsatT -> t mayUnsatT

convenience wrapper around check which propagate unsatifiability or checks satisfiability as required

val getAsg : t -> Vpl.Vector.Symbolic.t

Get an assignment of the variables which satisfies the input constraints. Calling getAsg is only meaningful on a value of type t returned by check or checkFromAdd.

val insertBack : Vpl.Var.t -> t -> t

val compl : t -> int -> t

val stricten : t -> int -> t mayUnsatT

val forget : t -> int -> t

val pr : (Vpl.Var.t -> string) -> t -> string

Pretty-print an element of type t. The function pr wraps functions prMat and prState.

val prMat : (Vpl.Var.t -> string) -> Vec.t option Vpl.Rtree.t -> string

prMat s pretty-prints the constraint matrix (field mat) of s.

val prState : (Vpl.Var.t -> string) -> state_t Vpl.Rtree.t -> string

prState s pretty-prints the bounds and values on variables (field state) of s.

val prBnd : bnd_t option -> string

Pretty-print an element of type bnd_t option.

val prSt : state_t -> string

Pretty-print an element of type state_t.