Module Vpl__UserInterface

Available polyhedral domains. There are three levels of certification:

No certification: no certificate is produced

OCaml certification: Each operator of the domain produces a certificate, which is checked by an OCaml procedure.

Coq certification: In addition to guarantees offered by OCaml certification, certificates are here checked by a Coq-proved procedure.

All domains fulfil a common interface (see Vpl__AbstractDomain.Type).

module UncertifiedQ : sig ... end

Uncertified domain on Q.

module UncertifiedZ : sig ... end

Uncertified domain on Z.

module OCamlCertifiedQ : sig ... end

OCaml certified domain on Q.

module OCamlCertifiedZ : sig ... end

OCaml certified domain on Z.

module CoqCertifiedQ : sig ... end

Coq certified domain on Q.

module CoqCertifiedZ : sig ... end

Coq certified domain on Z.