Module Vpl__Var

Types of variables, which are positive integers.

type t =

| XH
Last bit: represents value 1. The end of a path.

| XO of t
Bit 0. Take the left branch.

| XI of t
Bit 1. Take the right branch.

Type of variable

val u : t

First variable : equal to one.

val cmp : t -> t -> int

cmp x1 x2 compares x1 and x2 with "less" meaning "comes before in a breadth-first search". The result follows the same convention as Stdlib.compare.

val equal : t -> t -> bool

Equality test between two variables.

val toInt : t -> int

toInt p returns the integer which p represents, according to the Coq definition of positive. If the conversion of p overflows the int representation, Invalid_argument "Var.toInt" is raised.

val fromInt : int -> t

fromInt i converts a strictly positive integer into its bit string representation. If i is negative or nil, Invalid_argument "Var.fromInt" is raised.

val to_string : t -> string

Pretty-print a value v of type t as "vN", where N is the number denoted by v.

val to_string' : string -> t -> string

to_string' s v pretty-prints v as "sN", where N is the number denoted by v.

module Set : Stdlib.Set.S with type Set.elt = t

A module for manipulation sets of values of type t.

val horizon : Set.t -> t

horizon s returns the path which immediately follows the maximum path found in s.

val max : t list -> t

returns

the maximum of a given list of variables, w.r.t. cmp.

val next : t -> t

next v gives the path to the next variable in a breadth-first search in a tree. next of ... XO XH is ... XI XH, next of ... XO (XI XH) is ... XI (XO XH) and next of XI (XI XH) is XO (XO (XO XH)).

val of_prefixed_string : string -> t

Build a value of type t from a string of the form "xn", where x is a single letter and n is an integer. The resulting t has the value n.